Move popping current route from routing slip to ProcessCommand
Putting the responsibility for popping the current route into
the ProcessCommand, and creating an explicit interface for it
required by the Processor ends up giving more consistency in
the pipeline.
Now, the command source just provides the command as is.
The first service can add a routing slip, including itself
as part of the route.
The current route is popped off immediately before forwarding
the command.
Having an explicit interface for CloneWithoutThisRoute means
that next time I change my mind about where it goes, it can
be moved without much problem :-)