Remove ICommandFactory
The ICommandFactory was confusing. Move its functionality into the
router. Although this gives the router an additional responsibility
it makes the process flow clearer. The code implementing what
the ICommandFactory was doing can still be composed in.

Also, add a few comments into the main processor to help with
explanations.